Migrate cuml.experimental.fil -> cuml.fil (#6799)
Requires #6728 

Closes https://github.com/rapidsai/cuml/issues/6808 .

Authors:
  - Philip Hyunsu Cho (https://github.com/hcho3)
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6799
