+++ python/cuml/CMakeLists.txt
@@ -100,7 +100,7 @@ add_subdirectory(cuml/datasets)
 add_subdirectory(cuml/decomposition)
 add_subdirectory(cuml/ensemble)
 add_subdirectory(cuml/explainer)
-add_subdirectory(cuml/experimental/fil)
+add_subdirectory(cuml/fil)
 add_subdirectory(cuml/kernel_ridge)
 add_subdirectory(cuml/linear_model)
 add_subdirectory(cuml/manifold)
