--- python/cuml/cuml/experimental/fil/README.md
@@ -1,3 +1,0 @@
-# Experimental FIL - RAPIDS Forest Inference Library
-
-As of RAPIDS 25.04, experimental FIL has been promoted to stable. It is retained here temporarily to support users who have already migrated to experimental FIL, but it will be moved exclusively to `cuml.fil` in RAPIDS 25.06.
