+++ python/cuml/cuml/fil/CMakeLists.txt
@@ -25,5 +25,5 @@ rapids_cython_create_modules(
   CXX
   SOURCE_FILES "${cython_sources}"
   LINKED_LIBRARIES "${linked_libraries}"
-  MODULE_PREFIX experimental_fil_
+  MODULE_PREFIX fil_
 )
